requirements.txt
aiohttp == 3.7.4.post0
amqp == 5.0.6
anyjson == 0.3.3
asgiref == 3.4.1
astroid == 2.8.3
async-generator == 1.10
async-timeout == 3.0.1
attrs == 21.2.0
autopep8 == 1.5.7
Babel == 2.9.1
backoff == 1.10.0
backports.entry-points-selectable == 1.1.0
beautifulsoup4 == 4.9.1
billiard == 3.6.4.0
bleach == 4.1.0
boto == 2.49.0
boto3 == 1.9.88
botocore == 1.12.253
bump2version == 1.0.1
bumpversion == 0.6.0
celery == 5.1.2
certifi == 2021.10.8
cffi == 1.15.0
cfgv == 3.3.1
chardet == 4.0.0
charset-normalizer == 2.0.7
click == 7.1.2
click-didyoumean == 0.3.0
click-plugins == 1.1.1
click-repl == 0.2.0
colorama == 0.4.4
coverage == 4.2
cryptography == 3.3.2
cssselect == 1.1.0
decorator == 5.1.0
defusedxml == 0.7.1
dicttoxml == 1.7.4
diff-match-patch == 20200713
distlib == 0.3.3
Django == 3.2.8
django-admin-rangefilter == 0.6.2
django-cacheops == 6.0
django-cors-headers == 3.10.0
django-countries == 6.1.3
django-filter == 2.4.0
django-import-export == 2.3.0
django-json-widget == 1.0.1
django-next-prev == 1.0.1
django-phonenumber-field == 3.0.1
django-queryset-csv == 1.1.0
django-redis == 5.0.0
django-rest-passwordreset == 1.2.1
django-ses == 1.0.3
django-simple-history == 3.0.0
django-storages == 1.7.2
django-timezone-field == 4.1.2
django-webpack-loader == 0.6.0
djangorestframework == 3.12.2
djangorestframework-simplejwt == 4.3.0
docutils == 0.15.2
dparse == 0.5.1
drf-nested-routers == 0.91
et-xmlfile == 1.1.0
factory-boy == 2.7.0
fake-factory == 9999.9.9
Faker == 9.5.2
filelock == 3.3.1
flower == 0.9.3
freezegun == 0.3.15
funcy == 1.16
future == 0.18.2
geocoder == 1.38.1
geographiclib == 1.52
geoip2 == 4.4.0
geopy == 2.2.0
gunicorn == 20.1.0
h11 == 0.12.0
hiredis == 2.0.0
identify == 2.3.1
idna == 3.3
importlib-metadata == 4.8.1
ip2geotools == 0.1.5
IP2Location == 8.6.2
isort == 5.9.3
jeepney == 0.7.1
jmespath == 0.10.0
keyring == 23.2.1
kombu == 5.1.0
lazy-object-proxy == 1.6.0
lxml == 4.6.3
MarkupPy == 1.14
maxminddb == 2.2.0
mccabe == 0.6.1
meld3 == 2.0.1
mock == 2.0.0
multidict == 5.2.0
ndg-httpsclient == 0.5.1
nodeenv == 1.6.0
numpy == 1.21.3
nylas == 4.8.1
odfpy == 1.4.1
openpyxl == 3.0.9
outcome == 1.1.0
packaging == 21.0
pandas == 1.3.4
pbr == 5.6.0
phonenumbers == 8.10.16
Pillow == 8.4.0
pip == 21.3.1
pip-review == 1.1.0
pkginfo == 1.7.1
platformdirs == 2.4.0
pre-commit == 2.15.0
prompt-toolkit == 3.0.21
psycopg2-binary == 2.9.1
pusher == 2.1.4
pyasn1 == 0.4.8
pycodestyle == 2.8.0
pycparser == 2.20
pyfcm == 1.4.5
Pygments == 2.10.0
PyJWT == 1.7.1
pylint == 2.11.1
pymongo == 3.7.2
PyNaCl == 1.4.0
pynamodb == 3.3.3
pyOpenSSL == 21.0.0
pyparsing == 2.4.7
pyquery == 1.4.3
python-dateutil == 2.8.2
pytz == 2021.3
PyYAML == 6.0
ratelim == 0.1.6
readme-renderer == 30.0
redis == 3.5.3
requests == 2.26.0
requests-toolbelt == 0.9.1
rfc3986 == 1.5.0
s3transfer == 0.1.13
safety == 1.10.3
SecretStorage == 3.3.1
selenium == 3.141.0
sentry-sdk == 0.12.2
setuptools == 57.0.0
single-beat == 0.4.1
six == 1.16.0
sniffio == 1.2.0
sortedcontainers == 2.4.0
soupsieve == 2.2.1
sqlparse == 0.4.2
supervisor == 4.2.2
tablib == 3.0.0
text-unidecode == 1.3
toml == 0.10.2
tornadis == 0.8.1
tornado == 5.1.1
tqdm == 4.62.3
trio == 0.19.0
trio-websocket == 0.9.2
twilio == 6.38.1
twine == 3.4.2
typed-ast == 1.4.3
typing == 3.7.4.3
typing-extensions == 3.10.0.2
unicodecsv == 0.14.1
urllib3 == 1.25.11
URLObject == 2.4.3
vine == 5.0.0
virtualenv == 20.8.1
wcwidth == 0.2.5
webencodings == 0.5.1
wheel == 0.36.2
word2number == 1.1
wrapt == 1.13.2
wsproto == 1.0.0
xlrd == 2.0.1
XlsxWriter == 1.3.7
xlwt == 1.3.0
xmljson == 0.2.0
yarl == 1.7.0
zipp == 3.6.0
